prediction/requirements/base.txt
dicom_numpy==0.1.1
Flask==0.12.2
numpy==1.13.1
pydicom==0.9.9
keras==2.0.8
tensorflow==1.3.0
h5py==2.7.0
scipy==0.19.1
SimpleITK==1.0.1
